Create the user documentation site
We have decided to have public documentation for the cloud platform. We
want to make sure that this documentation is as easy to consume as
possible and also that teams can keep updated with progress by seeing
what has changed in the docs.

I decided to re-use the theme we have used for the [technical
guidance](1) site for consistency and to reduce the overhead of
working out something new.

(1): https://github.com/ministryofjustice/technical-guidance