Make attaching to a namespace explicit
We assume you are deploying your application into the namespace that you
created as part of the [creating an environment]() guide but we did not
make the commands explicit enough.

Here we've decided to add `--namespace` as the approach as it makes it
explicit what you are doing and is a bit quicker for now than
configuring a context in your .kubeconfig. Happy for this to be changed
in the future.