Fix incorrect link definition
The link definition that @alkar used to generate a relative link did not
work. I've made it more simple so that it does work.

I also changed the name of the page whilst I was at it as there will be
a few different pages about secrets so better to be more specific about
what we are talking about and made the expiry date for the page in the
future, rather than the past.