Changing the homepage layout: part 1
We've largely dumped the pages we've been creating into the
`cloud-platfrom` directory without too much thought. This is a proposal
for re-working the homepage so that the pages are more clearly laid out.

This is part 1. I am rearranging the pages around three thematic areas:

1. Getting started -- things that you need to do to get started on the
platform (like auth)
2. Deploying an app -- the process of deploying and app and then doing
things with it like adding secrets etc
3. Other topics -- anything that doesn't fit the first two for now. When
we get more of them we can re-order again.

This should make it easier to understand what is going on for a user.
The next parts are to move the pages around so that they fit into these
categories and order them on the page.