Change the homepage layout: part 2
Recategorise homepage and force ordering

**I've made NO content changes in this commit, I've only moved things
around**

As I said in the previous commit, the homepage needed to be rearranged
around the three areas: getting started, deploying an app and other
topics.

To make this more robust I gave each of the directories that make up
those sections a name starting with a number (e.g.
`01-getting-started`). This will make it easier to do ordering as jekyll
orders things off alpha-numeric ordering.

I then categorised each of the existing pages under one of those
headings. To order them within a heading (e.g. make "autheticate.md" the
first page to show up under the "Getting started" heading) I also added
numbers to the beginning of each of the files. That makes it easy to
change the ordering -- you just change the number to where you want that
page to appear (e.g. if you want it to be third add `003-` to the
beginning and make sure that the list goes, `001-, 002-, 003-, 004-` and
that you didn't add two `003s`.