Make relative links work
For relative links to work we need to use a something that interpolates
the base url of the site into the path, rather than just providing the
plain link. This is because you can get different base urls for your
site when you run the site in different places (e.g.
GitHub pages and locally).

This construction is a [jekyll
recommeded](https://jekyllrb.com/docs/github-pages/#project-page-url-structure) (and @stevemarshall
recommended) way of generating the relative links so that you can pick
up the right base url.