Update environment creation docs to use templates
As part of some prior work on adding network policies and resource
quotas @sablumiah created some templates for all of the resources we
want to define when we create a namespace. These live in
`namespace-resources`.

These templates would be useful for users to work from when creating
their namespaces. At the moment we do not have any information on how to
create a e.g. resource limit in a standard way. When we ask users to do
this we have to rely on them to parse the docs correctly and come up
with a good answer.

This commit changes the user documentation to guide the users through
using those templates when they create a new namespace. This means that
each time a user creates a new namespace using the docs they should get
a set of reasonably standard files with well thought out defaults.

I've also added a brief introduction to what each object that they are
defining does, and a link to the relevant Kubernetes documentation.