Update where to add rbac rules for Tiller
We have moved to telling users to put rbac policy in a `01-rbac.yaml`
file in their namespace. This page still told them to create a file of
the form `<service-env>-tiller-rbac.yaml` which is somewhat inconsistent
with the new naming pattern.

Instead we now ask them to append the rbac rules they need for Tiller to
their existing `01-rbac.yaml` file. This puts all rule based
authentication in one file which may or may not be a good idea in the
long run but gives clarity and consistency for now.