Add automatic environment creation to docs
@johuolape [has
added](https://github.com/ministryofjustice/cloud-platform-environments/pull/166)
functionality for us to automatically generate the namespace resource
files using terraform. He also did a draft of updating the guide to
reflect the change.

I've built on that to change the flow of the document a little and
remove some of the bits that were now overly wordly or incorrect. I've
tested out the method as well to make sure that this is resonably clear from a
user perspective.