Replace individual module docs with a generic one.
Having module documentated individually in this repository is a duplication and makes it hard to keep it up to date.
The documentation will now live in each module's repository and will be the source of truth. The new document introduced
in this commit acts as a list of the available modules along with some generic information on usage.