Make it clearer how to get started
We had a user request for "access" to the platform in the
`#cloud-platform-users` channel. This confused us as all you need to
access the platform is to be part of a Github team in the MoJ Github
organisation.

Turns out we had guidance that said people should ask in the channel to
get started. They don't need to do this. So I've made it clearer that
they can just get on with it if they have the right permissions and
instead added a section to say that they can get help and support
through the channel.